fix: Initialize root dependency graph
The original version version had the root web server be the root of the
dependency graph; and the message pump had been shoehorned in.

But the automatic verification of a valid graph didn't include the
message pump, which had broken during refactoring.

The fix is to create a new root graph object, containing both the web
server and the message pump.